+++ apps/backend/src/modules/users/libs/constants/constants.ts
@@ -1,3 +1,0 @@
-const MINIMAL_LENGTH = 0;
-
-export { MINIMAL_LENGTH };