+++ apps/backend/src/modules/users/libs/types/types.ts
@@ -1,5 +1,8 @@
+export { type UserTaskDay } from "./user-task-day.type.js";
 export {
 	type EmailDto,
+	type NotificationAnswersPayloadDto,
+	type NotificationAnswersRequestDto,
 	type ResetPasswordDto,
 	type UserDto,
 	type UserGetAllResponseDto,