+++ apps/backend/src/modules/users/user.controller.ts
@@ -13,11 +13,14 @@ import { userUpdateValidationSchema } from "~/modules/users/users.js";
 import { UsersApiPath } from "./libs/enums/enums.js";
 import { checkAccessToUserData } from "./libs/hooks/hooks.js";
 import {
+	type NotificationAnswersPayloadDto,
+	type NotificationAnswersRequestDto,
 	type UserDto,
 	type UserGetParametersDto,
 	type UserUpdateParametersDto,
 	type UserUpdateRequestDto,
 } from "./libs/types/types.js";
+import { notificationAnswersValidationSchema } from "./libs/validation-schemas/validation-schemas.js";
 
 /**
  * @swagger
@@ -41,7 +44,19 @@ import {
  *          updatedAt:
  *            type: string
  *            format: date-time
+ *      NotificationQuestionsRequest:
+ *        type: object
+ *        properties:
+ *          notificationFrequency:
+ *            type: string
+ *          userTaskDays:
+ *            type: array
+ *            items:
+ *              type: number
+ *          userId:
+ *            type: number
  */
+
 class UserController extends BaseController {
 	private userService: UserService;
 
@@ -66,6 +81,21 @@ class UserController extends BaseController {
 				),
 			method: "POST",
 			path: UsersApiPath.AVATAR,
+		});
+
+		this.addRoute({
+			handler: (options) =>
+				this.saveNotificationAnswers(
+					options as APIHandlerOptions<{
+						body: NotificationAnswersRequestDto;
+						user: UserDto;
+					}>,
+				),
+			method: "POST",
+			path: UsersApiPath.NOTIFICATION_QUESTIONS,
+			validation: {
+				body: notificationAnswersValidationSchema,
+			},
 		});
 
 		this.addRoute({
@@ -134,7 +164,7 @@ class UserController extends BaseController {
 	 *        - bearerAuth: []
 	 *      responses:
 	 *        200:
-	 *          description: Successfull operation
+	 *          description: Successful operation
 	 *          content:
 	 *            application/json:
 	 *              schema:
@@ -148,6 +178,45 @@ class UserController extends BaseController {
 	): Promise<APIHandlerResponse> {
 		return {
 			payload: await this.userService.find(options.params.id),
+			status: HTTPCode.OK,
+		};
+	}
+
+	/**
+	 * @swagger
+	 * /users/notification-questions:
+	 *    post:
+	 *      description: Save user preferences based on notification questions form
+	 *      security:
+	 *        - bearerAuth: []
+	 *      requestBody:
+	 *        required: true
+	 *        content:
+	 *          application/json:
+	 *            schema:
+	 *              $ref: "#/components/schemas/NotificationQuestionsRequest"
+	 *      responses:
+	 *        200:
+	 *          description: Successful operation
+	 *          content:
+	 *            application/json:
+	 *              schema:
+	 *                $ref: "#/components/schemas/User"
+	 */
+
+	private async saveNotificationAnswers(
+		options: APIHandlerOptions<{
+			body: NotificationAnswersPayloadDto;
+			user: UserDto;
+		}>,
+	): Promise<APIHandlerResponse> {
+		const updatedUserDto = await this.userService.saveNotificationAnswers(
+			options.user.id,
+			options.body,
+		);
+
+		return {
+			payload: updatedUserDto,
 			status: HTTPCode.OK,
 		};
 	}