+++ apps/backend/src/modules/users/user.entity.ts
@@ -1,4 +1,6 @@
-import { type Entity } from "~/libs/types/types.js";
+import { type Entity, type ValueOf } from "~/libs/types/types.js";
+
+import { NotificationFrequency } from "./libs/enums/enums.js";
 
 class UserEntity implements Entity {
 	private avatarFileId: null | number;
@@ -13,11 +15,15 @@ class UserEntity implements Entity {
 
 	private name: string;
 
+	private notificationFrequency: ValueOf<typeof NotificationFrequency>;
+
 	private passwordHash: string;
 
 	private passwordSalt: string;
 
 	private updatedAt: string;
+
+	private userTaskDays: null | number[];
 
 	private constructor({
 		avatarFileId,
@@ -26,9 +32,11 @@ class UserEntity implements Entity {
 		email,
 		id,
 		name,
+		notificationFrequency,
 		passwordHash,
 		passwordSalt,
 		updatedAt,
+		userTaskDays,
 	}: {
 		avatarFileId: null | number;
 		avatarUrl: null | string;
@@ -36,19 +44,23 @@ class UserEntity implements Entity {
 		email: string;
 		id: null | number;
 		name: string;
+		notificationFrequency: ValueOf<typeof NotificationFrequency>;
 		passwordHash: string;
 		passwordSalt: string;
 		updatedAt: string;
+		userTaskDays: null | number[];
 	}) {
 		this.avatarFileId = avatarFileId;
 		this.avatarUrl = avatarUrl;
 		this.createdAt = createdAt;
-		this.id = id;
 		this.email = email;
+		this.id = id;
 		this.name = name;
+		this.notificationFrequency = notificationFrequency;
 		this.passwordHash = passwordHash;
 		this.passwordSalt = passwordSalt;
 		this.updatedAt = updatedAt;
+		this.userTaskDays = userTaskDays;
 	}
 
 	public static initialize({
@@ -58,9 +70,11 @@ class UserEntity implements Entity {
 		email,
 		id,
 		name,
+		notificationFrequency,
 		passwordHash,
 		passwordSalt,
 		updatedAt,
+		userTaskDays,
 	}: {
 		avatarFileId: null | number;
 		avatarUrl: null | string;
@@ -68,9 +82,11 @@ class UserEntity implements Entity {
 		email: string;
 		id: null | number;
 		name: string;
+		notificationFrequency?: ValueOf<typeof NotificationFrequency>;
 		passwordHash: string;
 		passwordSalt: string;
 		updatedAt: string;
+		userTaskDays?: number[];
 	}): UserEntity {
 		return new UserEntity({
 			avatarFileId,
@@ -79,9 +95,12 @@ class UserEntity implements Entity {
 			email,
 			id,
 			name,
+			notificationFrequency:
+				notificationFrequency ?? NotificationFrequency.NONE,
 			passwordHash,
 			passwordSalt,
 			updatedAt,
+			userTaskDays: userTaskDays ?? null,
 		});
 	}
 
@@ -103,9 +122,11 @@ class UserEntity implements Entity {
 			email,
 			id: null,
 			name,
+			notificationFrequency: NotificationFrequency.NONE,
 			passwordHash,
 			passwordSalt,
 			updatedAt: "",
+			userTaskDays: null,
 		});
 	}
 
@@ -115,9 +136,11 @@ class UserEntity implements Entity {
 		createdAt: string;
 		email: string;
 		name: string;
+		notificationFrequency: ValueOf<typeof NotificationFrequency>;
 		passwordHash: string;
 		passwordSalt: string;
 		updatedAt: string;
+		userTaskDays: number[];
 	} {
 		return {
 			avatarFileId: this.avatarFileId,
@@ -125,9 +148,11 @@ class UserEntity implements Entity {
 			createdAt: this.createdAt,
 			email: this.email,
 			name: this.name,
+			notificationFrequency: this.notificationFrequency,
 			passwordHash: this.passwordHash,
 			passwordSalt: this.passwordSalt,
 			updatedAt: this.updatedAt,
+			userTaskDays: this.userTaskDays ?? [],
 		};
 	}
 
@@ -138,7 +163,9 @@ class UserEntity implements Entity {
 		email: string;
 		id: number;
 		name: string;
+		notificationFrequency: ValueOf<typeof NotificationFrequency>;
 		updatedAt: string;
+		userTaskDays: number[];
 	} {
 		return {
 			avatarFileId: this.avatarFileId,
@@ -147,7 +174,9 @@ class UserEntity implements Entity {
 			email: this.email,
 			id: this.id as number,
 			name: this.name,
+			notificationFrequency: this.notificationFrequency,
 			updatedAt: this.updatedAt,
+			userTaskDays: this.userTaskDays ?? [],
 		};
 	}
 }