+++ apps/backend/src/modules/users/user.repository.ts
@@ -4,9 +4,9 @@ import { type FileModel } from "~/modules/files/files.model.js";
 import { UserEntity } from "~/modules/users/user.entity.js";
 import { type UserModel } from "~/modules/users/user.model.js";
 
-import { MINIMAL_LENGTH } from "./libs/constants/constants.js";
-import { type UserUpdateRequestDto } from "./libs/types/types.js";
+import { type UserTaskDay } from "./libs/types/types.js";
 import { type UserDetailsModel } from "./user-details.model.js";
+import { type UserTaskDaysModel } from "./user-task-days.model.js";
 
 class UserRepository implements Repository {
 	private fileModel: typeof FileModel;
@@ -15,13 +15,22 @@ class UserRepository implements Repository {
 
 	private userModel: typeof UserModel;
 
-	public constructor(
-		userModel: typeof UserModel,
-		userDetailsModel: typeof UserDetailsModel,
-		fileModel: typeof FileModel,
-	) {
+	private userTaskDaysModel: typeof UserTaskDaysModel;
+
+	public constructor({
+		fileModel,
+		userDetailsModel,
+		userModel,
+		userTaskDaysModel,
+	}: {
+		fileModel: typeof FileModel;
+		userDetailsModel: typeof UserDetailsModel;
+		userModel: typeof UserModel;
+		userTaskDaysModel: typeof UserTaskDaysModel;
+	}) {
 		this.userModel = userModel;
 		this.userDetailsModel = userDetailsModel;
+		this.userTaskDaysModel = userTaskDaysModel;
 		this.fileModel = fileModel;
 	}
 
@@ -63,14 +72,27 @@ class UserRepository implements Repository {
 		});
 	}
 
-	public delete(): ReturnType<Repository["delete"]> {
-		return Promise.resolve(true);
+	public async delete(id: number): Promise<boolean> {
+		const rowsDeleted = await this.userModel.query().deleteById(id);
+
+		return Boolean(rowsDeleted);
+	}
+
+	public async deleteUserTaskDays(userId: number): Promise<boolean> {
+		const rowsDeleted = await this.userTaskDaysModel
+			.query()
+			.delete()
+			.where({ userId });
+
+		return Boolean(rowsDeleted);
 	}
 
 	public async find(id: number): Promise<null | UserEntity> {
 		const user = await this.userModel
 			.query()
-			.withGraphFetched(RelationName.USER_DETAILS)
+			.withGraphFetched(
+				`[${RelationName.USER_DETAILS}, ${RelationName.USER_TASK_DAYS}]`,
+			)
 			.findById(id);
 
 		const avatar = user?.userDetails.avatarFileId
@@ -89,9 +111,13 @@ class UserRepository implements Repository {
 					email: user.email,
 					id: user.id,
 					name: user.userDetails.name,
+					notificationFrequency: user.notificationFrequency,
 					passwordHash: user.passwordHash,
 					passwordSalt: user.passwordSalt,
 					updatedAt: user.updatedAt,
+					userTaskDays: user.userTaskDays.map(
+						(taskDay: UserTaskDay) => taskDay.dayOfWeek,
+					),
 				})
 			: null;
 	}
@@ -99,7 +125,9 @@ class UserRepository implements Repository {
 	public async findAll(): Promise<UserEntity[]> {
 		const users = await this.userModel
 			.query()
-			.withGraphFetched(RelationName.USER_DETAILS);
+			.withGraphFetched(
+				`[${RelationName.USER_DETAILS}, ${RelationName.USER_TASK_DAYS}]`,
+			);
 
 		return await Promise.all(
 			users.map(async (user) => {
@@ -114,9 +142,13 @@ class UserRepository implements Repository {
 					email: user.email,
 					id: user.id,
 					name: user.userDetails.name,
+					notificationFrequency: user.notificationFrequency,
 					passwordHash: user.passwordHash,
 					passwordSalt: user.passwordSalt,
 					updatedAt: user.updatedAt,
+					userTaskDays: user.userTaskDays.map(
+						(taskDay: UserTaskDay) => taskDay.dayOfWeek,
+					),
 				});
 			}),
 		);
@@ -126,8 +158,10 @@ class UserRepository implements Repository {
 		const user = await this.userModel
 			.query()
 			.where({ email })
-			.withGraphFetched(RelationName.USER_DETAILS)
-			.first();
+			.first()
+			.withGraphFetched(
+				`[${RelationName.USER_DETAILS}, ${RelationName.USER_TASK_DAYS}]`,
+			);
 
 		if (!user) {
 			return null;
@@ -152,41 +186,37 @@ class UserRepository implements Repository {
 
 	public async update(
 		id: number,
-		payload: Partial<UserUpdateRequestDto>,
+		payload: Partial<UserDetailsModel>,
 	): Promise<null | UserEntity> {
-		const { name } = payload;
-
-		const userDetailsUpdateData: Partial<UserDetailsModel> = {
-			...(name && { name }),
-		};
-
-		if (Object.keys(userDetailsUpdateData).length > MINIMAL_LENGTH) {
-			await this.userDetailsModel
-				.query()
-				.patch(userDetailsUpdateData)
-				.where({ userId: id });
-		}
-
+		const userDetails = await this.userDetailsModel
+			.query()
+			.findOne({ userId: id });
+		const updatedUserDetails = await userDetails
+			?.$query()
+			.patchAndFetch(payload);
 		const user = await this.userModel
 			.query()
 			.findById(id)
-			.withGraphFetched(RelationName.USER_DETAILS);
+			.withGraphFetched(RelationName.USER_TASK_DAYS);
+		const avatar = await this.fileModel
+			.query()
+			.findById(updatedUserDetails?.avatarFileId as number);
 
-		const avatar = user?.userDetails.avatarFileId
-			? await this.fileModel.query().findById(user.userDetails.avatarFileId)
-			: null;
-
-		return user
+		return user && updatedUserDetails
 			? UserEntity.initialize({
-					avatarFileId: avatar?.id ?? null,
+					avatarFileId: updatedUserDetails.avatarFileId,
 					avatarUrl: avatar?.url ?? null,
 					createdAt: user.createdAt,
 					email: user.email,
 					id: user.id,
-					name: user.userDetails.name,
+					name: updatedUserDetails.name,
+					notificationFrequency: updatedUserDetails.notificationFrequency,
 					passwordHash: user.passwordHash,
 					passwordSalt: user.passwordSalt,
 					updatedAt: user.updatedAt,
+					userTaskDays: user.userTaskDays.map(
+						(taskDay: UserTaskDay) => taskDay.dayOfWeek,
+					),
 				})
 			: null;
 	}
@@ -199,15 +229,19 @@ class UserRepository implements Repository {
 			.query()
 			.findOne({ url: payload.url });
 
-		await this.userDetailsModel
+		const userDetails = (await this.userDetailsModel
 			.query()
 			.patch({ avatarFileId: avatarFile?.id as number })
-			.where({ userId: id });
+			.where({ userId: id })
+			.returning("*")
+			.first()) as UserDetailsModel;
 
 		const user = await this.userModel
 			.query()
 			.findById(id)
-			.withGraphFetched(RelationName.USER_DETAILS);
+			.withGraphFetched(
+				`[${RelationName.USER_DETAILS}, ${RelationName.USER_TASK_DAYS}]`,
+			);
 
 		const avatar = user?.userDetails.avatarFileId
 			? await this.fileModel.query().findById(user.userDetails.avatarFileId)
@@ -221,9 +255,13 @@ class UserRepository implements Repository {
 					email: user.email,
 					id: user.id,
 					name: user.userDetails.name,
+					notificationFrequency: userDetails.notificationFrequency,
 					passwordHash: user.passwordHash,
 					passwordSalt: user.passwordSalt,
 					updatedAt: user.updatedAt,
+					userTaskDays: user.userTaskDays.map(
+						(taskDay: UserTaskDay) => taskDay.dayOfWeek,
+					),
 				})
 			: null;
 	}
@@ -252,6 +290,19 @@ class UserRepository implements Repository {
 			passwordSalt: user.passwordSalt,
 			updatedAt: user.updatedAt,
 		});
+	}
+
+	public async updateUserTaskDays(
+		id: number,
+		userTaskDays: number[],
+	): Promise<void> {
+		await this.deleteUserTaskDays(id);
+
+		await this.userTaskDaysModel.query().insert(
+			userTaskDays.map((day) => {
+				return { dayOfWeek: day, userId: id };
+			}),
+		);
 	}
 }