+++ apps/backend/src/modules/users/user.service.ts
@@ -13,6 +13,7 @@ import { type UserRepository } from "~/modules/users/user.repository.js";
 
 import { UserError } from "./libs/exceptions/exceptions.js";
 import {
+	type NotificationAnswersPayloadDto,
 	type UserDto,
 	type UserGetAllResponseDto,
 	type UserSignUpRequestDto,
@@ -81,6 +82,17 @@ class UserService implements Service {
 
 	public findByEmail(email: string): Promise<null | UserEntity> {
 		return this.userRepository.findByEmail(email);
+	}
+
+	public async saveNotificationAnswers(
+		id: number,
+		payload: NotificationAnswersPayloadDto,
+	): Promise<null | UserEntity> {
+		await this.userRepository.updateUserTaskDays(id, payload.userTaskDays);
+
+		return await this.userRepository.update(id, {
+			notificationFrequency: payload.notificationFrequency,
+		});
 	}
 
 	public async update(