+++ apps/frontend/src/libs/components/balance-wheel-chart/libs/constants/animation-interval.constant.ts
@@ -1,3 +1,0 @@
-const ANIMATION_INTERVAL = 1000;
-
-export { ANIMATION_INTERVAL };