+++ apps/frontend/src/libs/components/balance-wheel-chart/libs/enums/dot-offset-x.enum.ts
@@ -1,0 +1,9 @@
+const DotOffsetX = {
+	LEFT_BOTTOM: 7,
+	LEFT_UPPER: 14,
+	LEFTMOST: 2,
+	RIGHTMOST: 14,
+	UPPER_CENTER: 5,
+} as const;
+
+export { DotOffsetX };