+++ apps/frontend/src/libs/components/balance-wheel-chart/libs/helpers/get-absolute-center/get-absolute-center.helper.ts
@@ -1,0 +1,7 @@
+import { AngleCoefficient } from "../../enums/enums.js";
+
+const getAbsoluteCenter = (lowerBound: number, upperBound: number): number => {
+	return lowerBound + (upperBound - lowerBound) / AngleCoefficient.HALF;
+};
+
+export { getAbsoluteCenter };