+++ apps/frontend/src/libs/components/balance-wheel-chart/libs/types/polar-area.type.ts
@@ -1,0 +1,3 @@
+type PolarAreaType = "polarArea";
+
+export { type PolarAreaType };