+++ apps/frontend/src/modules/quiz/quiz-api.ts
@@ -2,7 +2,10 @@ import { APIPath, ContentType } from "~/libs/enums/enums.js";
 import { BaseHTTPApi } from "~/libs/modules/api/api.js";
 import { type HTTP } from "~/libs/modules/http/http.js";
 import { type Storage } from "~/libs/modules/storage/storage.js";
-import { type QuizQuestionDto } from "~/modules/quiz/quiz.js";
+import {
+	type QuizQuestionDto,
+	type QuizScoresGetAllResponseDto,
+} from "~/modules/quiz/quiz.js";
 
 import { QuizApiPath } from "./libs/enums/enums.js";
 
@@ -28,6 +31,19 @@ class QuizApi extends BaseHTTPApi {
 		);
 
 		return await response.json<{ items: QuizQuestionDto[][] }>();
+	}
+
+	public async getScores(): Promise<QuizScoresGetAllResponseDto> {
+		const response = await this.load(
+			this.getFullEndpoint(QuizApiPath.SCORE, {}),
+			{
+				contentType: ContentType.JSON,
+				hasAuth: true,
+				method: "GET",
+			},
+		);
+
+		return await response.json<QuizScoresGetAllResponseDto>();
 	}
 }